test: add reusable ResizeObserver mock
